addin crud annotation and mapping control layer

+++ src/main/java/fr/spring/eBoutique/project/control/ILigneCommandeControl.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface ILigneCommandeControl {
 
-	List<LigneCommande> getLignesCommande(Integer idCommande);
+	List<LigneCommande> getLignesCommande(Long idCommande);
 	LigneCommande addLigneCommande(LigneCommande ligneCommande);
 	
 }
